+++ src/App.tsx
@@ -7,14 +7,19 @@ import { Provider } from "react-redux";
 import { Route, Switch } from "react-router-dom";
 import { ConnectedRouter } from "react-router-redux";
 import { ThroughProvider } from "react-through";
+import { sleep } from "lib/util";
 import { history, default as store } from "./configureStore";
 import * as css from "./layouts/App.scss";
 
-export class App extends React.Component<{}, {arcIsInitialized: boolean}> {
+export class App extends React.Component<{}, {
+  arcIsInitialized: boolean;
+  retryingArc: boolean;
+}> {
   constructor(props: {}) {
     super(props);
     this.state = {
       arcIsInitialized: false,
+      retryingArc: false,
     };
   }
 
@@ -23,10 +28,16 @@ export class App extends React.Component<{}, {arcIsInitialized: boolean}> {
     // not create a provider for the app, rather will just initialize Arc with a
     // readonly provider with no account, internal only to it.
     initializeArc()
-      .then((): void => {
+      .then(async (success: boolean) => {
+        while (!success) {
+          this.setState({ retryingArc: true });
+          await sleep(5000);
+          success = await initializeArc();
+        }
         this.setState({ arcIsInitialized: true });
       })
       .catch ((err): void => {
+        // eslint-disable-next-line no-console
         console.log(err);
       });
 
@@ -48,9 +59,16 @@ export class App extends React.Component<{}, {arcIsInitialized: boolean}> {
     });
   }
 
-  public render(): any {
+  public render(): RenderOutput {
     if (!this.state.arcIsInitialized) {
-      return <div className={css.loading}><Loading/></div>;
+      return (
+        <div className={css.waitingToInitContainer}>
+          { this.state.retryingArc ? 
+            <div className={css.waitingToInitMessage}>Waiting to connect to the blockchain.  If this is taking a while, please ensure that you have a good internet connection.</div> : ""
+          }
+          <div className={css.loading}><Loading/></div>
+        </div>
+      );
     } else  {
       return (
         <Provider store={store}>