+++ src/components/Account/AccountProfileName.tsx
@@ -17,10 +17,11 @@ interface IProps {
 
 export default class AccountProfileName extends React.Component<IProps, null> {
 
-  public render() {
+  public render(): RenderOutput {
     const { accountAddress, accountProfile, daoAvatarAddress, historyView, detailView } = this.props;
 
     const accountNameClass = classNames({
+      [css.accountName]: true,
       [css.detailView]: detailView,
       [css.historyView]: historyView,
     });