+++ src/components/Proposal/ActionButton.tsx
@@ -1,21 +1,16 @@
-import { Address, IDAOState, IProposalStage, IProposalState, IRewardState, Reward } from "@daostack/client";
+import { Address, IDAOState, IProposalOutcome, IProposalStage, IProposalState, IRewardState } from "@daostack/client";
 import { executeProposal, redeemProposal } from "actions/arcActions";
 import { enableWeb3ProviderAndWarn } from "arc";
 import * as classNames from "classnames";
 import { ActionTypes, default as PreTransactionModal } from "components/Shared/PreTransactionModal";
-import withSubscription, { ISubscriptionProps } from "components/Shared/withSubscription";
 import { claimableContributionRewards, hasClaimableRewards } from "lib/util";
-import * as moment from "moment";
 import Tooltip from "rc-tooltip";
 import * as React from "react";
 import { connect } from "react-redux";
 import { IRootState } from "reducers";
-import { closingTime } from "reducers/arcReducer";
 import { proposalEnded } from "reducers/arcReducer";
 import { showNotification } from "reducers/notifications";
 import { IProfileState } from "reducers/profilesReducer";
-import { Observable, of} from "rxjs";
-import { map, mergeMap } from "rxjs/operators";
 import * as css from "./ActionButton.scss";
 /* import RedemptionsString from "./RedemptionsString"; */
 import RedemptionsTip from "./RedemptionsTip";
@@ -24,10 +19,13 @@ import BN = require("bn.js");
 
 interface IExternalProps {
   currentAccountAddress?: Address;
-  dao: IDAOState;
+  daoState: IDAOState;
   daoEthBalance: BN;
   detailView?: boolean;
+  expanded?: boolean;
   proposalState: IProposalState;
+  rewards: IRewardState;
+  expired: boolean;
 }
 
 interface IStateProps {
@@ -40,7 +38,7 @@ interface IDispatchProps {
   showNotification: typeof showNotification;
 }
 
-type IProps = IExternalProps & IStateProps & IDispatchProps & ISubscriptionProps<IRewardState>;
+type IProps = IExternalProps & IStateProps & IDispatchProps;
 
 const mapStateToProps = (state: IRootState, ownProps: IExternalProps): IExternalProps & IStateProps => {
   const proposalState = ownProps.proposalState;
@@ -71,7 +69,7 @@ class ActionButton extends React.Component<IProps, IState> {
 
   public async handleClickExecute(_event: any): Promise<void> {
     if (!(await enableWeb3ProviderAndWarn(this.props.showNotification.bind(this)))) { return; }
-    await this.props.executeProposal(this.props.dao.address, this.props.proposalState.id, this.props.currentAccountAddress);
+    await this.props.executeProposal(this.props.daoState.address, this.props.proposalState.id, this.props.currentAccountAddress);
   }
 
   public async handleClickRedeem(_event: any): Promise<void> {
@@ -83,24 +81,24 @@ class ActionButton extends React.Component<IProps, IState> {
     this.setState({ preRedeemModalOpen: false });
   }
 
-  public render(): any {
-    const rewardsForCurrentUser = this.props.data;
-
+  public render(): RenderOutput {
     const {
       beneficiaryProfile,
       currentAccountAddress,
-      dao,
+      daoState,
       daoEthBalance,
       detailView,
+      expired,
+      expanded,
       proposalState,
+      rewards,
     } = this.props;
 
     const executable = proposalEnded(proposalState) && !proposalState.executedAt;
-    const expired = closingTime(proposalState).isSameOrBefore(moment());
 
     let beneficiaryHasRewards;
 
-    const accountHasGPRewards = hasClaimableRewards(rewardsForCurrentUser);
+    const accountHasGPRewards = hasClaimableRewards(rewards);
     if (proposalState.contributionReward) {
       const daoBalances: {[key: string]: BN} = {
         eth: daoEthBalance,
@@ -113,14 +111,17 @@ class ActionButton extends React.Component<IProps, IState> {
       beneficiaryHasRewards = Object.keys(contributionRewards).length > 0;
     }
 
-    const redeemable = currentAccountAddress ? proposalState.executedAt && proposalState.accountsWithUnclaimedRewards.includes(currentAccountAddress.toLowerCase()) : false;
+    const redeemable = proposalState.executedAt &&
+                       ((currentAccountAddress ? proposalState.accountsWithUnclaimedRewards.includes(currentAccountAddress.toLowerCase()) : false)
+                        || (proposalState.winningOutcome === IProposalOutcome.Pass && beneficiaryHasRewards));
+
     // hack to work around https://github.com/daostack/subgraph/issues/304
     const redemptionsTip = RedemptionsTip({
       beneficiaryHasRewards,
       currentAccountAddress,
-      dao,
+      dao: daoState,
       proposal: proposalState,
-      rewardsForCurrentUser,
+      rewardsForCurrentUser: rewards,
     });
 
     const redeemButtonClass = classNames({
@@ -130,6 +131,7 @@ class ActionButton extends React.Component<IProps, IState> {
     const wrapperClass = classNames({
       [css.wrapper]: true,
       [css.detailView]: detailView,
+      [css.expanded]: expanded,
     });
 
     return (
@@ -140,7 +142,7 @@ class ActionButton extends React.Component<IProps, IState> {
             action={this.handleRedeemProposal}
             beneficiaryProfile={beneficiaryProfile}
             closeAction={this.closePreRedeemModal.bind(this)}
-            dao={dao}
+            dao={daoState}
             effectText={redemptionsTip}
             proposal={proposalState}
           /> : ""
@@ -171,7 +173,7 @@ class ActionButton extends React.Component<IProps, IState> {
                 : redeemable ?
                   <div>
                     {/* !detailView ?
-                  <RedemptionsString currentAccountAddress={currentAccountAddress} dao={dao} proposal={proposalState} rewards={rewardsForCurrentUser} />
+                  <RedemptionsString currentAccountAddress={currentAccountAddress} dao={daoState} proposal={proposalState} rewards={rewardsForCurrentUser} />
                   : ""
               */}
                     <Tooltip placement="left" trigger={["hover"]} overlay={redemptionsTip}>
@@ -204,33 +206,13 @@ class ActionButton extends React.Component<IProps, IState> {
 
     const {
       currentAccountAddress,
-      dao,
+      daoState,
       proposalState,
       redeemProposal,
     } = this.props;
 
-    await redeemProposal(dao.address, proposalState.id, currentAccountAddress);
+    await redeemProposal(daoState.address, proposalState.id, currentAccountAddress);
   }
 }
 
-const SubscribedActionButton = withSubscription({
-  wrappedComponent: ActionButton,
-  loadingComponent: <div>Loading...</div>,
-  errorComponent: (props) => <div>{ props.error.message }</div>,
-
-  checkForUpdate: (oldProps, newProps) => { return oldProps.proposalState.id !== newProps.proposalState.id || oldProps.currentAccountAddress !== newProps.currentAccountAddress; },
-
-  createObservable: (props: IProps) => {
-    const proposalState = props.proposalState;
-
-    if (props.currentAccountAddress) {
-      return proposalState.proposal.rewards({ where: {beneficiary: props.currentAccountAddress}})
-        .pipe(map((rewards: Reward[]): Reward => rewards.length === 1 && rewards[0] || null))
-        .pipe(mergeMap(((reward): Observable<IRewardState> => reward ? reward.state() : of(null))));
-    } else {
-      return of(null);
-    }
-  },
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(SubscribedActionButton);
+export default connect(mapStateToProps, mapDispatchToProps)(ActionButton);