+++ src/components/Proposal/ProposalSummary/ProposalSummarySchemeRegistrar.tsx
@@ -32,7 +32,7 @@ export default class ProposalSummary extends React.Component<IProps, IState> {
     this.setState({ network: (await getNetworkName()).toLowerCase() });
   }
 
-  public render(): any {
+  public render(): RenderOutput {
     const { proposal, detailView, transactionModal } = this.props;
 
     const proposalSummaryClass = classNames({
@@ -119,6 +119,9 @@ export default class ProposalSummary extends React.Component<IProps, IState> {
                           {
                             // eslint-disable-next-line no-bitwise
                             permissions & 16 ? <div>Call genericCall on behalf of</div> : ""
+                          }
+                          {
+                            <div>Mint or burn reputation</div>
                           }
                         </td>
                       </tr>