+++ src/components/Shared/PreTransactionModal.tsx
@@ -11,7 +11,6 @@ import { getExchangesList, humanProposalTitle } from "lib/util";
 import Tooltip from "rc-tooltip";
 import * as React from "react";
 import { connect } from "react-redux";
-//@ts-ignore
 import { Modal } from "react-router-modal";
 import { showNotification } from "reducers/notifications";
 import { IProfileState } from "reducers/profilesReducer";
@@ -78,7 +77,7 @@ class PreTransactionModal extends React.Component<IProps, IState> {
     this.setState({ instructionsOpen: !this.state.instructionsOpen });
   }
 
-  public render() {
+  public render(): RenderOutput {
     const { actionType, beneficiaryProfile, currentAccount, currentAccountGens, dao, effectText, proposal, secondaryHeader } = this.props;
     const { stakeAmount } = this.state;