+++ CHANGELOG.md
@@ -7,9 +7,15 @@
 ### Unreleased
 
   - Features Added
-    - ..
+    - Links in proposal description open in new tab
+    - Show a message on startup when waiting for subgraph
+    - Create a single place where all redemptions can be redeemed
   - Bugs Fixed
-    - ..
+    - Align headers of table in proposal history
+    - Re-add redeem for beneficiary button to proposal details page
+    - Fix error on initializing arc
+    - Fix infinite scroll loading of queued scheme proposals
+
 
 
 ### 2019-09-12 [actually two releses]
@@ -27,12 +33,19 @@
     - improvements, bug fixes on the Scheme Information page
     - DAO Discussion page
     - Add link to help center
+    - Reduce number of subscriptions from DAO history page
+    - add breadcrums to mobile
+    - add support for WalletConnect web3 providers
+    - add static mint and burn permission 
   - Bugs Fixed
     - Fix number formatting
     - Fix crashing of scheme page
     - Hide notifications after 10 seconds
     - Add dao address as default value for Cross-DAO redemptions
     - Fix redemptions count
+    - Display proper message when no history
+    - avoid duplicate Genesis Alpha cards
+    - fix action button not appearing when proposal card countdown completes
 
 
 ### version 0.9.4; 2019-08-13